Add `Integer` type annotation to `sizehint!` overloads
JuliaLang/julia#59168 changed the base signature of `sizehint` to use
`::Integer` instead of `::Any` for the second argument, thus packages
also overload against the signature instead in order to avoid method
ambiguity error.

+++ src/little_dict.jl
@@ -196,7 +196,7 @@ end
 
 ######## Methods that all mutable AbstractDict's should implement
 
-function Base.sizehint!(dd::UnfrozenLittleDict, sz)
+function Base.sizehint!(dd::UnfrozenLittleDict, sz::Integer)
     sizehint!(dd.keys, sz)
     sizehint!(dd.vals,sz)
     return dd
